+++ resources/lang/en/item_status.php
@@ -4,4 +4,5 @@
         'not_ordered' => 'Not ordered',
         'delivered' => 'Delivered',
         'unknown' => 'Unknown',
+        'partially_delivered' => 'Partially delivered',
     ];